move response processor into seperate file to be ts-ignored
